get rid of fabric importing config-api
